+++ src/main/java/jp/co/sony/ppog/service/SbsSsmCrudLogicService.java
@@ -36,7 +36,7 @@ public interface SbsSsmCrudLogicService {
 	 * @param id 都市ID
 	 * @return List<String>
 	 */
-	List<String> getListOfNationsById(Long id);
+	List<String> getListOfNationsById(Integer id);
 
 	/**
 	 * 入力した都市情報を保存する
@@ -57,7 +57,7 @@ public interface SbsSsmCrudLogicService {
 	 *
 	 * @param id 都市ID
 	 */
-	void removeById(Long id);
+	void removeById(Integer id);
 
 	/**
 	 * 大陸情報を取得する
